+++ README.md
@@ -21,8 +21,8 @@
   - [x] 게임에 참여할 사람은 각각 이름을 갖는다.
 - [ ] 카드의 합이 21을 넘지 않을 경우 얼마든지 카드를 계속 뽑을 수 있다.
   - [ ] 패자가 가려질 경우 자동으로 게임을 종료한다.
-  - [ ] 그 외의 경우 모든 플레이어가 카드 뽑기를 멈출 때까지 게임을 진행한다.
+  - [x] 그 외의 경우 모든 플레이어가 카드 뽑기를 멈출 때까지 게임을 진행한다.
 - [x] 카드의 숫자는 카드 숫자를 기본으로 한다.
   - [x] King, Queen, Jack은 10이다.
   - [ ] Ace는 1 또는 11이 될 수 있다.
-- [ ] 게임에 참여할 사람의 이름을 입력할 수 있다.
+- [x] 게임에 참여할 사람의 이름을 입력할 수 있다.
