+++ src/main/kotlin/blackjack/domain/participant/PlayerName.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 
 @JvmInline
 value class PlayerName(
