+++ src/test/kotlin/blackjack/domain/participant/PlayerTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 
 import blackjack.domain.deck.Card
 import blackjack.domain.deck.CardNumber.ACE
